+++ gazebo/gui/ModelAlign.cc
@@ -172,7 +172,8 @@ void ModelAlign::AlignVisuals(std::vector<rendering::VisualPtr> _visuals,
       if (it->first)
       {
         it->first->SetWorldPose(it->second);
-        it->first->SetTransparency(it->first->GetTransparency()*2.0 - 1.0);
+        it->first->SetTransparency(std::abs(
+            it->first->GetTransparency()*2.0-1.0));
       }
     }
     this->dataPtr->originalVisualPose.clear();