+++ gazebo/gui/building/WallInspectorDialog.cc
@@ -172,13 +172,13 @@ WallInspectorDialog::WallInspectorDialog(QWidget *_parent)
   this->textureComboBox->setMinimumWidth(50);
   this->textureComboBox->setMinimumHeight(50);
   this->textureComboBox->setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Fixed);
-  this->textureList.push_back(":/images/wood.png");
-  this->textureList.push_back(":/images/ceiling_tiled.png");
-  this->textureList.push_back(":/images/bricks.png");
+  this->textureList.push_back(":wood.jpg");
+  this->textureList.push_back(":tiles.jpg");
+  this->textureList.push_back(":bricks.png");
   for (unsigned int i = 0; i < this->textureList.size(); ++i)
   {
-    this->textureComboBox->addItem(QPixmap(this->textureList[i]),
-        QString(""));
+    this->textureComboBox->addItem(QPixmap(this->textureList[i]).scaled(
+        QSize(30, 30), Qt::IgnoreAspectRatio), QString(""));
   }
   this->textureComboBox->addItem("X");
   this->textureComboBox->setCurrentIndex(this->textureComboBox->count()-1);
@@ -310,14 +310,17 @@ void WallInspectorDialog::SetThickness(double _thickness)
 /////////////////////////////////////////////////
 void WallInspectorDialog::SetColor(const QColor _color)
 {
+  // Find index corresponding to color (only a few colors allowed so far)
+  int index = 0;
   for (unsigned int i = 0; i < this->colorList.size(); ++i)
   {
     if (this->colorList[i] == _color)
     {
-      this->colorComboBox->setCurrentIndex(i);
-      return;
+      index = i;
+      break;
     }
   }
+  this->colorComboBox->setCurrentIndex(index);
 
   // Add a new color
   this->colorList.push_back(_color);